data-engineering/requirements.txt
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.1
